% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{class}
\name{knapsack-class}
\alias{knapsack-class}
\alias{knapsack}
\title{knapsack problem}
\description{
This is package of 'knapsack' problem.
}
\section{Fields}{

\describe{
\item{\code{x}}{A data frame.}

\item{\code{W}}{A numeric.}

\item{\code{item}}{A vector.}
}}


