% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{data.farme}
\usage{
knapsack_objects
}
\description{
This knapsack_objects is for the test of lab6.
}
\keyword{datasets}
